Use USM Host space for USMObjectMem instead of shared
This results in a significant speedup. However, there's a good chance that this change is superseded entirely by https://github.com/kokkos/kokkos/pull/4097, so in the end this is unlikely to be needed.

I would like to come back and properly template the SYCL memory spaces to avoid repetitive code, but right now it's not a priority.